Read WWK CSV files

Function to return WWK data retrieved in csv format as a tidy tibble.

Usage

wwk_read_csv(file, skip = 2, n_max = Inf, show_col_types = FALSE)

wwk_read_csv_indicator(file, skip = 2, n_max = Inf, show_col_types = FALSE)

Arguments

file

Either a path to a file, a connection, or literal data (either a single string or a raw vector).

Files ending in .gz, .bz2, .xz, or .zip will be automatically uncompressed. Files starting with http://, https://, ftp://, or ftps:// will be automatically downloaded. Remote gz files can also be automatically downloaded and decompressed.

Literal data is most useful for examples and tests. To be recognised as literal data, the input must be either wrapped with I(), be a string containing at least one new line, or be a vector containing at least one string with a new line.

Using a value of clipboard() will read from the system clipboard.

skip

Number of lines to skip before reading data. If comment is supplied any commented lines are ignored after skipping.

n_max

Maximum number of lines to read.

show_col_types

If FALSE, do not show the guessed column types. If TRUE always show the column types, even if they are supplied. If NULL (the default) only show the column types if they are not explicitly supplied by the col_types argument.

Value

a tibble

Details

The CSV files returned by the WWK API cannot be used directly for further analyses, as they contain:

Lines with metadata at the beginning and end of the data set.

The first line contains the names of the indicators.

The second line contains the names of the regions.

Additional meta data is provided at the end.

The data is returned in wide format:

Each row corresponds to an indicator.

The first column contains the name of the indicator.

All further columns contain the data for a region-year pair.

Examples

if (FALSE) { # \dontrun{
csv <- wwk_export(
  indikator = c("geburten", "sterbefaelle"),
  region = c("koeln", "muenchen")
)
wwk_read_csv(csv)
} # }